Move data utilities to @serverless/utils
