+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "serverless-components",
   "version": "0.1.2",
-  "description": "PoC for the next version of the Serverless Framework",
+  "description": "An easier way to build applications with cloud services",
   "author": "Serverless, Inc.",
   "license": "Apache-2.0",
   "main": "src/index.js",
@@ -29,19 +29,16 @@
   },
   "dependencies": {
     "@octokit/rest": "^14.0.5",
-    "@serverless/utils": "0.0.1",
+    "@serverless/utils": "0.0.2",
     "analytics-node": "^3.2.0",
-    "archiver": "^2.1.1",
     "aws-sdk": "^2.214.1",
     "axios": "^0.17.1",
     "bluebird": "^3.5.1",
     "chalk": "^2.3.2",
     "concordance": "^3.0.0",
-    "crypto-extra": "^0.4.0",
     "fs-extra": "^5.0.0",
     "graphlib": "^2.1.5",
     "js-yaml": "^3.10.0",
-    "mime-types": "^2.1.18",
     "minimist": "^1.2.0",
     "node-fetch": "^2.1.1",
     "node-where": "^1.1.0",
@@ -49,14 +46,10 @@
     "ramda": "^0.25.0",
     "raml-validate": "^1.1.2",
     "raven": "^2.4.2",
-    "rc": "^1.2.6",
     "s3-client": "^4.4.1",
-    "semver": "^5.5.0",
-    "uuid": "^3.2.1",
-    "write-file-atomic": "^2.3.0"
+    "semver": "^5.5.0"
   },
   "devDependencies": {
-    "adm-zip": "^0.4.9",
     "babel-jest": "^22.4.1",
     "eslint": "^4.18.1",
     "eslint-config-airbnb-base": "^12.1.0",
