+++ registry/rest-api/package.json
@@ -13,6 +13,7 @@
   "author": "serverless.com",
   "license": "Apache-2.0",
   "dependencies": {
+    "@serverless/utils": "0.0.2",
     "ramda": "^0.25.0"
   }
 }
