+++ registry/s3-dirloader/package.json
@@ -13,7 +13,7 @@
   "author": "serverless.com",
   "license": "Apache-2.0",
   "dependencies": {
-    "@serverless/utils": "0.0.1",
+    "@serverless/utils": "0.0.2",
     "aws-sdk": "^2.201.0",
     "mime-types": "^2.1.18",
     "ramda": "^0.25.0"
