+++ src/utils/components/generateContext.js
@@ -5,9 +5,9 @@ const getInstanceId = require('./getInstanceId')
 const getChildrenPromises = require('./getChildrenPromises')
 const getComponentFunctions = require('./getComponentFunctions')
 const getComponentRootPath = require('./getComponentRootPath')
+const log = require('../logging/log')
 const getServiceId = require('../state/getServiceId')
 const getState = require('../state/getState')
-const log = require('../log')
 
 const generateContext = (
   components,
