--- src/utils/getRegistryRoot.js
@@ -1,3 +1,0 @@
-const path = require('path')
-
-module.exports = () => path.join(__dirname, '..', '..', 'registry')
