+++ src/utils/index.js
@@ -1,27 +1,19 @@
-const misc = require('./misc')
-const log = require('./log')
-const getRegistryRoot = require('./getRegistryRoot')
-const state = require('./state')
-const telemetry = require('./telemetry')
 const components = require('./components')
 const dag = require('./dag')
-const mapIndexed = require('./mapIndexed')
-const reduceIndexed = require('./reduceIndexed')
-const reduceObjIndexed = require('./reduceObjIndexed')
+const logging = require('./logging')
+const misc = require('./misc')
+const registry = require('./registry')
+const state = require('./state')
+const telemetry = require('./telemetry')
 const variables = require('./variables')
-const validateCoreVersion = require('./validateCoreVersion')
 
 module.exports = {
   ...components,
   ...dag,
+  ...logging,
   ...misc,
+  ...registry,
   ...state,
   ...telemetry,
-  ...variables,
-  getRegistryRoot,
-  validateCoreVersion,
-  log,
-  mapIndexed,
-  reduceIndexed,
-  reduceObjIndexed
+  ...variables
 }
