+++ src/utils/logging/index.js
@@ -1,0 +1,5 @@
+const log = require('./log')
+
+module.exports = {
+  log
+}
