--- src/utils/mapIndexed.js
@@ -1,8 +1,0 @@
-const {
-  addIndex,
-  map
-} = require('ramda')
-
-const mapIndexed = addIndex(map)
-
-module.exports = mapIndexed
