--- src/utils/misc/forEachIndexed.js
@@ -1,5 +1,0 @@
-const { addIndex, forEach } = require('ramda')
-
-const forEachIndexed = addIndex(forEach)
-
-module.exports = forEachIndexed
