+++ src/utils/misc/index.js
@@ -1,9 +1,5 @@
-const forEachIndexed = require('./forEachIndexed')
-const getRandomId = require('./getRandomId')
 const handleSignalEvents = require('./handleSignalEvents')
 
 module.exports = {
-  forEachIndexed,
-  getRandomId,
   handleSignalEvents
 }
