--- src/utils/reduceIndexed.js
@@ -1,8 +1,0 @@
-const {
-  addIndex,
-  reduce
-} = require('ramda')
-
-const reduceIndexed = addIndex(reduce)
-
-module.exports = reduceIndexed
