+++ src/utils/registry/getRegistryRoot.js
@@ -1,0 +1,3 @@
+const path = require('path')
+
+module.exports = () => path.join(__dirname, '..', '..', '..', 'registry')
