+++ src/utils/registry/index.js
@@ -1,0 +1,5 @@
+const getRegistryRoot = require('./getRegistryRoot')
+
+module.exports = {
+  getRegistryRoot
+}
